spec_exec: fix can_be_ignored function
Previous implementation did not make much sense. We would prematurely return
from speculative execution for some errors which could be ignored.

I adjusted the function, so we use `#[deny(clippy::wildcard_enum_match_arm)]`
lint. It will force the developer to think before adjusting this function
if new QueryError variant is introduced.

I categorized the errors (see the comments in the code) and justified
why each of them needs to be ignored or not.