frame/value: extract legacy module
To make deprecations in value.rs less invasive, as measured by the
number of required changes in the next commit, a `legacy` module is
extracted for all legacy items.
As a bonus, the future deletion of legacy items will be clearer.

I recommend viewing this commit with whitespace changes ignored.