support serialization flattening for references
If a struct's field is a reference, allow that reference to use the
`flatten` attribute so that the field can be serialized using
SerializeRow and the columns produced flattened into the parent
struct.

Prior to this commit this was not possible so this commit does two
changes:

1. Adds a blanket implementation on  the needed
traits (SerializeRowByName and SerializeRowInOrder) for references of
structs that already implement them.

2. Adds a bound to the lifetimes of the fields when making the partial
structs during serialization such that the reference to the fields
have to outlive the partial struct being created during
serialization.