topology: De-duplicate query_tables_schema calls
This function performs a request that fetches columns for all tables
and views. It is potentially the most performance-impactful part of the
schema fetching process, and it was unnecessarily called twice:
in query_tables and in query_views.

It can be easily prevented by calling the function earlier and passing
the result to query_tables and query_views.