metadata: convert errors in query_filter_keyspace_name
I introduced a utility trait which converts multiple lower-level errors
to a specific error type that provides more context.

The function is now generic over the trait implementations.

After this commit, there is an outdated comment next to nested
`make_keyspace_filtered_query_pager` function. It will be addressed
in the following commit.