errors: introduce UseKeyspaceError
Note: I wasn't able to split it into multiple commits. It's always tricky
to split the changes into multiple commits when there are some tokio channels
we use to send a `Result<_, SomeError>` where SomeError is the error type we
adjust.

Major changes:
- UseKeyspaceProtocolError is now renamed to UseKeyspaceError. New variants are introduced.
- Removed ProtocolError::UseKeyspace variant
- Added UseKeyspaceError variant to QueryError and NewSessionError
- Replaced `QueryError` with `UseKeyspacError` in USE KEYSPACE execution path
  in `Session`, `Cluster`, `ClusterWorker`, `Node`, `PoolRefiller` and `Connection` methods.

There is a potential FIXME I introduced. Pay attention to this during review.