session: pass custom timestamp in request execution APIs
This is a preparation commit for introducing new methods on Session that
will allow setting custom timestamps on statements upon executing them.

For now, the user-facing API still does not allow setting a custom
timestamp, but the internal functions do pass the optional timestamp.

The timestamp still does not reach the connection layer, but this will be
adjusted in a next commit.