statement: remove per-statement custom timestamp setting
A timestamp is something related to a single execution of a statement,
not to the statement itself. Also, setting a timestamp on a statement
inside the Batch would not set a timestamp on a Batch itself, which has
shown to be a pitfall.
This commit finishes the migration to the new per-execution timestamp
API by removing timestamp settings from statements.