Extract UserDefinedType from ColumnType
As part of an effort to unify CqlType with ColumnType we need to make
those types compatible enough. CqlType stored UserDefinedType as a
separate struct under Arc, in order to share it between different
metadata structures. This commit does the same for ColumnType.

Note that the field for UDT name is now "name" instead of the "type_name",
because the old one was a bit redundant.