scylla: rename "ssl" feature to "openssl-010"
As openssl crate is still before its 1.0 (stable) release, we need to
prepare for its possible introduces breaking changes. As it's present in
our public API, we hide it behind a versioned feature flag. This way,
if in the future openssl is released in 0.11 or 1.0 version, we'll
simply add the new feature flag and gradually deprecate the old one.

As a side note, openssl crate hasn't issued a major release for 7 years
now, so it's plausible that it's not going to happen (at least soon).

Co-authored-by: Wojciech Przytuła <wojciech.przytula@scylladb.com>