pager: narrow return error type of `QueryPager::new_for_*`
Last time during pager API errors refactor I missed that this return error
type can be narrowed.

There are two kinds of callers:
1. Session::query/execute_iter
In this case, we needed to replace `QueryError::NextRowError` variant
with NextPageError. However, this variant will be removed at the end of this PR as
we introduce a new error type for aforementioned Session methods.

2. Connection::query/execute_iter
This is an internal API. It's only used during metadata fetch. I decided to
map it to `NextRowError`, because it doesn't make much sense to distinguish between these two
from user's perspective. NextRowError will be returned later anyway, when we iterate over the rows.