integration: `for_each_target_execute` supports nonsharded clusters
Before, `for_each_target_execute` would panic on absence of Sharder
for a Node. This could be caused by two main reasons:
1) the node was unsharded (e.g., a Cassandra node),
2) there were no open connections to the node, e.g. due to the node
   being down.

In both cases, we prefer not to panic in `for_each_target_execute` but
rather assume the node is unsharded and proceed. In the case 1),
execution of the request should succeed, whereas in 2) we should get
the ConnectionPoolError.