integration: add test that queries all nodes directly
I noticed that we lacked a test that requires direct connectivity to all
nodes in the cluster in order to pass. Other tests would often succeed
even if only one node (the initial contact point) was directly
reachable.

An example of a test which has exercised direct connectivity to all
nodes is tablets.rs, yet it could be enabled only for non-cloud case,
as it uses the proxy.