errors: rename and extract TracingError
It was previously called `TracingProtocolError` and was contained in ExecutionError::ProtocolError.

Since we now have a big picture of errors hierarchy, it's much clearer that `TracingError` should
be a separate error type returned by tracing methods. ExecutionError should NOT depend on it.

This commit renamed TracingProtocolError to TracingError, and changes the return type
of Session::get_tracing_info(). New variant is introduced to `TracingError` - it contains
an `ExecutionError`, which makes sense since we try to execute queries to system tracing
tables.