errors: inline RequestAttemptError in ExecutionError
This commit inlines the RequestAttemptError in ExecutionError.

It also adjusts some tests which otherwise are failing. This is because
some variants land in other place than they would before this commit.

In the next commits we will start cleaning up the ExecutionError - i.e.
removing unused variants and moving some of them to other places.

The purpose why we inline RequestAttemptError is to remove some
error variants from the top level. For example, `CqlErrorParseError` should
not be placed on the top level of the error hierarchy. It is a very specific
and users should get some more context about it - i.e. that this is the cause
of failure of the last attempt.