partitioner: expose PartitionerName
There are two reasons to do so.
First, users may be interested which partitioner is used for a specific
PreparedStatement.
Second, tests in session_test.rs call `get_partitioner_name()` to assert
some things. If we want to extract `session_test.rs` to integration
tests directory - out of `scylla` crate - we must therefore expose
`get_partitioner_name().

NOTE: PartitionerHasherAny is made `pub` in this commit. This is a
temporary measure. This is needed because otherwise the compiler errors
out that we leak a crate-private type in public interface (the
associated type in `impl Partitioner for PartitionerName`).
We will be able to revert this once we unpub `Partitioner(Hasher)`.