proxy: don't use COMPRESSED flag in tests randomly
Some tests just set random CQL flags and assert that they are simply
passed through the proxy. It used to work, but as now we make the proxy
clever enough to interpret at least one of them, namely the compression
flag, we need to stop abusing it blindly.

The compression flag is the bit 0x1, so we simply must not use any odd
numbers as flags.