column_specs: introduce additional lifetime parameter
This is to distinguish between the lifetime of the slice, and the lifetime
of column spec. It matters when someone creates ColumnSpecs from some slice
of `ColumnSpec<'static>` (e.g. PreparedStatement).

NOTE: this commit introduces lifetime bound `'slice: 'spec` which is required
only because `TableSpecView` and `ColumnSpecView` are parameterized by single lifetime.
This bound will be removed once we remove the aforementioned view structs - the bound
is obviously not true if `'spec == 'static` (case for `PreparedStatement`).