statement: rename query.rs -> statement.rs
This results in existence of `scylla::statement::statement` module. I don't like
it, so I unpubbed the inner module (previously query), and re-exported
`Statement` type in the outer module. I'm not sure if this is the best approach,
and I don't have a strong opinion on this. I added a TODO in the code - it lists
three (in my opinion) most reasonable solutions to this problem.

Currently, users can access statement types like this:
- scystatement::Statement
- scylla::prepared_statement::PreparedStatement (or scylla::statement::prepared_statement::[...])
- scylla::batch::Batch (or scylla::statement::batch::[...])