partitioner.rs: un-pub calculate_token_for_partition_key
This function operates on SerializedValues, and so is not type-safe.
Given that we already provide token-calculation API on PreparedStatement
and ClusterState, exposing this method seems redundant.

If it turns out that there are users who need this method and can't use
existing APIs then we can think about providing appropriate safe API,
or making this pub once again as a last resort.