CI: Remove verbose output
It has not been useful even once in my experience, but it makes output
unreadable.
If we ever need it, we can restore it.

I wanted to make it conditional. The way to do this is to enable the env
variable that controls verbosity if the CI is run in debug mode.
Github Actions allows checking it with "runner.debug" variable.
Obvious idea is to put something like
`CARGO_TERM_VERBOSE: "${{ runner.debug == '1' && 'true' || 'false' }}"`
in the workflow's "env" section.

This does not work, because `runner` context is not available in workflow
env, and not even in job env. This means we would have to put this
env in every single step that may need verbose, which is absurd.

Another example of GitHub being a truly terrible tool.