Cassandra: Remove build step
This step does not bring us any benefit. It also makes the CI run longer,
because the project has to be recompiled in the next step anyway,
because it has different RUSTFLAGS.

Instead of changing the RUSTFLAGS in the build step (and risking similar
problems in the future), let's remove it.