Table: Add Vec of primary key column specs
We want to move token calculation methods in ClusterState to accept
SerializeRow. This in turn means we need to serialize those values,
so we have to create RowSerializationContext from the Table struct.

RowSerializationContext currently needs a slice of ColumnSpec.
Table has no such slice. Instead it has a hashmap from column name to
a ColumnSpec, and a Vec of primary key column names.

We have three options:
- Add a field with the required slice to Table struct.
- Modify RowSerializationContext somehow so it can be created from the
  data that we already have in Table. I'm not sure how to do that, idea
  would be appreciated.
- Hybrid: Modify both Table and RowSerializationContext to make them
  work together.

This commit takes the first approach because it seemed to be the easiest
one. Doing it a different way is of course open for discussion.

Co-authored-by: Wojciech Przytuła <wojciech.przytula@scylladb.com>