Metadata: Verify that there are no gaps in pk and ck
Previously it was possible that some positions of `partition_key` and
`clustering_key` would remain unfilled and thus empty strings.

The probability was very low - Scylla would have to return very weird
data - but the possibility was there.

This commit verifies that this is not happening, and returns and error
if it is.

Co-authored-by: Wojciech Przytuła <wojciech.przytula@scylladb.com>