Move macro docs to scylla-macros crate
Given that we want scylla-macros to be strongly tied to scylla-cql it
makes sense to remove docs of the macros from the `scylla` crate.

Docs were moved as-is. The only change is fixing the links to other
items. Because `scylla` and `scylla-cql` are not in dependencies of
`scylla-macros`, it is not possible to link to those items as before.
Instead we need to use relative link to html file. This trick is also
used by e.g. Tokio.

One disadvantage of this is that the re-export of those macros need to
be at the fixed module depth in order for those relative links to work.
For example, we can't re-export the macro both as crateA::macroX and
crateB::macros:macroX. Further commits will fix the re-export to adhere
to this restriction.

Despite this issue (which one could argue is not an issue because it
forces better consistency) the change is a net positive. Before it, the
macro docs were rendered only in the `scylla` crate docs. Now they will
be rendered in all 3 crates (`scylla`, `scylla-cql`, `scylla-macros`)
and the links to other items will work in first two.