De-duplicate paths for macros
Previously users of `scylla` crate could import macros either using
`scylla::macroX` or `scylla::macros::macroX`.
We generally want to avoid having items in two different paths. Here it
is even more important, because of macro doc comments which require
macro re-exports to be at a set depth in order for links in doc comments
to work.

The main question is which path to leave. For now I decided to use
`scylla::macroX` path. Why?
- I checked few other Rust libraries (Tokio, Clap, rustyline, tracing,
  serde). As far as I can tell, they all re-export macros this way
  instead of putting them in some module. I don't know if there is any
  good reason for that, or if that is just a convention that libs seem
  to agree on, but either way it is imo a good idea to follow them.
- Looking through our examples, and some different projects using the
  driver, I see `scylla::macroX` used more often than
  `scylla::macros::macroX`.

I'm open to changing this if there are convincing arguments to use
`scylla::macros::macroX` path.

Two small related issues fixed in this commit:
- Doc comments for deserialization imported macros from scylla_cql
  instead of scylla.
- lib.rs in scylla and scylla-cql used wildcard, which is less explicit
  and more error-prone.