state: introduce new getters for keyspace info
Thanks to that, we will be able to remove the getter that returns a HashMap<String, Keyspace>.
Because of that, we will be able to change the underlying structure of how we store keyspace info
in the future. One idea is to potentially wrap `Keyspace` struct in an `Arc`, so we may possibly
share the keyspace info in multiple places.