scylla_macros: Fix broken links
Previous links worked locally for both scylla and scylla-cql, but on
docs.rs they only worked for scylla-cql. This is because on docs.rs you
have only docs for a single crate, so we can't refer to scylla-cql from
scylla. Fortunately we don't need to refer to it, because scylla
re-export all of those items.