Integration tests: Add structure for CCM integration
ccm module will contain tests that require ccm.
It's lib submodule will contain the ccm integration.
Why do it this way - which is different than what we did during a
hackathon?
- Old way required ugly hacks to share test utils between integration
test targets, and those hacks did not work well with rust-analyzer.
- One target means better compilation time

CCM tests will be guardded by a cfg, so we will still be able to run the
subset that we want:
- All tests: run integration tests with the required cfg
- Only CCM tests: as above, but filter by ccm folder
- Only non-ccm test: run without the cfg