config: introduce WriteCoalescingDelay
Introduced `WriteCoalescingDelay` enum and exposed configuration options in `SessionConfig`
and `SessionBuilder`.

Adjusted the documentation.

Notice that ideally, the `enable_write_coalescing` and newly introduced `write_coalescing_delay`
should be merged into one. Probably as `Option<WriteCoalescingDelay>`. We can't do it now,
as it would break the API. I'll open an issue regarding this after PR is merged (and put it under 2.0 milestone).