connection: pass event_sender and compression directly to reader
Thanks to that, we can prevent cloning `write_coalescing_delay` from the config.
Currently, the `WriteCoalescingDelay` enum is cheaply clonable (copiable),
but it may not be in the future.

In addition, (IMO) it makes the code a bit clearer - we explicitly pass
the payload from the config which is needed by the reader.