state/metadata: get and expose cluster version
In cpp-driver, `cass_schema_meta_version` returns the output
of `release_version` column from `system.local` query executed on
current control connection.

I'm not entirely sure if we should expose such API. `ClusterState::get_nodes_info()`
would be enough to implement in on the cpp-driver side - we would just choose
the version of random/first node. Or maybe return an error if there is a version
mismatch (?). This can be discussed.

The main thing I hate about this commit is this ugly `fold` expression
in `query_peers_and_cluster_version`. I have no idea if there is any
cleaner solution to this.