fix flipped errors when deriving SerializeRow
`NoColumnWithName` and `ValueMissingForColumn` were
flipped.

`NoColumnWithName` is returned when there is a serializable field in
the Rust struct (i.e., is it not ignored) but there is no matching
column for it in the spec

`ValueMissingForColumn` should be returned when the spec expects a
column but there is no matching field in the Rust struct.