session: prepare: introduce fallback to all shards
Before this commit, we could unnecessarily return an error from prepare:
Let's say we have a 1-node cluster, and we have a broken connection, and
we don't retry on another - we would retry error to the user despite
possibly being able to prepare.

This commit introduces fallback logic to `Session::prepare()`:
if preparation on a single (random) connection to every node fails,
the whole preparation is retried, this time on a single connection to
every shard.