connection: introduce `Connection::prepare_raw()`
`Connection::prepare_raw()` extracts majority of the logic from
`Connection::prepare()`, but it returns `RawPreparedStatement` instead
of `PreparedStatement`, delaying construction of `PreparedStatement` to
the call site. This way it's the caller's choice to materialise
`PreparedStatement` or not.
`Connection::prepare()` has the same behaviour as before; it calls
`Connection::prepare_raw()` and then materialises the
`PreparedStatement`.