codewide: introduce Coordinator to QueryResult
`QueryResult` now contains `Option<Coordinator>`. For this,
`NonErrorQueryResponse::into_query_result()` now accepts `Coordinator`.
It is constructed in `Session::run_request()` for each target from
`Plan` that is tried.

Note that `QueryResult::request_coordinator()` return `&Coordinator`
even though `QueryResult` may possibly hold `None`. If `None` is there,
the driver panics upon the getter called. Rationale: some private
`Connection` APIs (namely, `query_raw_unpaged()`) return `QueryResult`
even though they don't have access to `Node`, which is required to
construct a `Coordinator`. Even if we redesigned this particular API
to be able to create a `Coordinator`, we would still be left with
the fundamental chicken-and-egg problem that occurs when we use
the `Connection` APIs on the control connection to issue the initial
metadata fetch. In such case we have 0 chance to have the `Node`
accessible (as it's only constructed based on the first metadata fetch).

To solve the problem, those APIs construct `QueryResult` using a
dedicated constructor: `new_with_unknown_coordinator()`.
It should be used only in APIs that do not leak `QueryResult` to the
user, so the panic is not triggered.