CachingSession: Allow skipping result metadata
As `CachingSession` manages preparation internally, it does not expose
its `PreparedStatement`s in any way. Therefore, it's not possible for a
user of `CachingSession` to call
`PreparedStatement::set_use_cached_result_metadata` in order to take
advantage of that optimisation.

There are use cases when not having possibility to turn the optimisation
on may result in significant inefficiency. As an example, the
experimental NodeJS-over-Rust driver's API suits `CachingSession`'s API
best, but lack of the mentioned optimisation makes the experimental
driver fall short compared to the original NodeJS driver (which, most
likely, has the optimisation turned on).

Why did I decide to expose a `set_*` method instead of constructor?
Because there are already two constructors, so I'd have to add two more,
which would look quite unpleasant (and naming would be difficult). It
would be even more of a problem if we add more parameters in the future.

Having a setter is also more flexible: user can change the parameter
after the session creation.