CachingSession: Allow skipping result metadata
As `CachingSession` manages preparation internally, it does not expose
its `PreparedStatement`s in any way. Therefore, it's not possible for a
user of `CachingSession` to call
`PreparedStatement::set_use_cached_result_metadata` in order to take
advantage of that optimisation.

There are use cases when not having possibility to turn the optimisation
on may result in significant inefficiency. As an example, the
experimental NodeJS-over-Rust driver's API suits `CachingSession`'s API
best, but lack of the mentioned optimisation makes the experimental
driver fall short compared to the original NodeJS driver (which, most
likely, has the optimisation turned on).

New method is added to recently introduced CachingSessionBuilder,
allowing user to enable caching result metadata.