connection_pool: manually implement (Ref)UnwindSafe
These implementations are a temporary solution to the following problem:
`QueryResult` used to implement `(Ref)UnwindSafe`, but then we wanted it
to store a reference to `Node`. This, transitively, made it store
`NodeConnectionPool`, which did not implement those traits.
Thus, they would no longer be auto-implemented for `QueryResult`,
breaking the public API. Not to introduce an API breakage in a minor
release, we decided to manually hint that `NodeConnectionPool` is indeed
unwind-safe. Even if our we are wrong and the hint is misleading,
the documentation of those traits, not being `unsafe` traits, considers
them merely guidelines, not strong guarantees.