caching_session: builder: warn about cache capacity
The legacy way to construct `CachingSession` (the `from()` method)
required providing cache capacity explicitly. The builder has
some default capacity and no longer requires explicit configuration
of it. To protect users from the pitfall of having cache that is
too small for their needs, a warning is added to the builder's
docstring.