integration: fix test_fetch_system_keyspace()
Having read the contents of the test, I was confused what it actually
tests. `git blame` shed some light, so I pasted the commit message as
a doc comment: "This test case indicates that we support enough CQL
types to parse schema keyspace information.". A-ha!

But we no longer type check nor deserialize rows eagerly!
The deserialization refactor seems to have rendered this test case
useless. To fix this, I materialise all rows as Row, which actually
performs the type check and deserialization. This way I virtually
bring back the old, intended semantics of this test.