CachingSession: Store Arc<Session>
There is nothing that would prevent this since Session is designed for
concurrent usage, and this change will allow reusing a single Session
for multiple CachingSession (and to be used outside them as well). This
fits into our recommendation of only having a single Session.

Caching metadata parameter in CachingSession that I introduced in
previous commits is another rationale for this: user may want to execute
some requests with metadata caching and some without. Creating two
Sessions for that is wasteful.