Require schema agreement response from DDL coordinator
This fixes the issue described in parent commit.

Internal schema agreement APIs now accept a `Option<Uuid>` that is a
host id of a node that must be a part of schema agreement. For
user-requested agreements it will be None, and for agreements after DDL
it will be the coordinator.