Replace allow with expect where possible
If some check always (independent of configuration) triggers for some
code, it is better to use expect because we will be notified when the
check stops triggering.

Some uses of "allow" remain. We could try to replace them with
`cfg_attr` magic, but it would be compplicated in many cases and I feel
that it won't bring much benefit.