Clippy: Check also pub items
Turns out clippy has a terrible and shocking default: it won't make
suggestions that require introducing breaking changes.
This means that e.g. if we introduce a type which has a naming that
clippy would normally complain about, it won't if the type is pub!

For this reason we have many such types. This commit disables this
absurd default, and adds comment and attribute to each place affected by
this.