Cargo.toml: Enable all features for docs.rs
This is needed for docs.rs to include all features in the generated
documentation. It seems that `scylla` crate has already been configured
correctly, but `scylla-cql`, `scylla-macros`, and `scylla-proxy` were
missing the needed setting. Also, I added quotes around the `docs.rs`
metadata key to disambiguate it from the period that is used for
nesting metadata keys.