session_builder: make openssl-dependent doctest cfg-feature
The doc test for the `tls_context` method in `SessionBuilder` would fail
when the `openssl-010` feature was disabled, because the feature
was required to use the `openssl` crate required by the example.
The changes are introduces as suggested here:
https://users.rust-lang.org/t/doctests-that-require-a-non-default-feature-is-it-possible/29529/3
This makes all doc tests pass both when the `openssl-010` feature is
enabled or disabled.

Fixes: #1369