errors: improve EmptyPlan error message
As noted in #747, before the error refactor the `EmptyPlan` error
message had not been very helpful for users of the `DefaultPolicy`.
Even after the error refactor, the message still did not provide
hint to check the preferred datacenter when using `DefaultPolicy`.
This commit improves the error message to mention this particular
possible cause of the error.

Fixes: #747