statement: introduce BoundStatement
BoundStatement represents a prepared statement that has already been
bound with values to be executed with. It is purposely a lightweight
pair of both, with no separate config held. For now, there's no way to
configure the BoundStatement after its creation. It only inherits the
configuration from the PreparedStatement it's built from.

Co-Authored-By: Andres Medina <andresnrx@gmail.com>