default LBP: rate limit sanity checks warnings
In case of misconfiguration, the default load balancing policy can throw
a lot of warnings, which flood the logs and overwhelm the machine.
To prevent this, we rate limit the warnings there to one per second.

NOTE: each warning is rate limited separately, so if there are multiple
misconfigurations, each will still be logged.