session: de-genericise run_request(_speculative_fiber)
In the past, those functions needed to be generic over the result type
ResT, because they actually were instantiated with different result
types. Now, they are always instantiated with the same type:
NonErrorQueryResponse. This means we can de-genericise both those
functions for better readability and maintainability.

As a bonus, we can remove the trait `AllowedRunRequestResTType`.