deserialize: Fix rust name in type check for sets
Impls for BTreeSet and HashSet had a bug: they performed type check on
element, but then renamed the rust name in error to container type
(while cql type remained the element type). Such errors are obviously
incorrect.
There are 2 ways to fix it. I'll let the reviewer choos which we should
use:
- Performing type check on element, then wrapping the error (this is the
same thing that ListLikeIterator does).
- Calling type check of ListLikeIterator - shorter and prettier code,
but performs the same check twice (probably would be optimized out by
compiler). I chose this solution for now because it is consistent with
Vec and avoids code duplication.