crates: allow clippy::uninlined_format_args
clippy has started to complain about format strings that are not
inlined. In order to satisfy clippy's demands, we would have to
introduce a lot of changes. At the same time, I don't see any value
in avoiding use of format strings that are not inlined.

Thus, to calm clippy down, we allow this lint in all crates.