SE: No new fibers after exhausted plan
Execution plan is shared between fibers. That means if one fiber
declares the plan was exhausted, any new spawned fiber will do the same.
Because of that, there is no need to spawn new fibers in this situation.
The way to do that is setting remaining speculative retries to 0.
We can't return immediately - we still have to wait for remaining
running fibers.