DeserializeRow: default_when_null attribute
By default, if the DB provides a null value for a column when the Rust
type expects a non-null value, deserialization fails. Instead, if
`default_when_null` is specified for a field, deserialization yields
`Default::default()` if the column's value is null. This is important
for production as when adding columns to a table, existing rows will
have the new field filled with null. Clients can specify
`default_when_null` to handle that situation without using `Option`

Fixes: https://github.com/scylladb/scylla-rust-driver/issues/1399