Introduce safe_format helper
Itertools::format panics on double format. They could solve this by
introducing a Clone bound on iterator, but this would of course be a
breaking change. Writing such improved format is easy, so we can just do
this ourselves. This commit introduces extensions trait that exposes
safe_format method on cloneable iterators. Implementation internally
defers to Itertools::format for each formatting, to make code simpler.

I had to duplicate this between scylla and scylla-cql. I don't see a
problem with this - it is a single very small and simple utility that
has logic independent from the driver.