tree: fix module docs include path
Upon releasing scylla 1.3.1, a problem arose: the include path passed as
an argument to `include_str!` in the `scylla/src/lib.rs` file was
only correct when building the project with the workspace set up in the
following way, i.e., so that the `scylla-cql` crate is a sibling of the
`scylla` crate in the directory tree:
```
---- scylla-cql
 |-- scylla
```
However, when packaging the `scylla` crate for distribution, it is placed
in a directory with names that do not match the above, i.e.:
```
target/package ---- scylla-cql-<version>
                |-- scylla-<version>
```

We decided to fix this by adding a symlink in the `scylla/src/deserialize`
directory that points to the `scylla-cql/src/deserialize/README.md` file,
so that the `include_str!` macro can find it regardless of the directory
structure. `cargo publish` will then copy the symlink into the package,
ensuring that the documentation is still accessible when building the
package.